+++ src/main/java/com/banquito/core/loans/DTO/TiposComisionDTO.java
@@ -15,5 +15,5 @@ public class TiposComisionDTO {
     private String tipoCalculo;
     private BigDecimal monto;
     private String estado;
-    private BigDecimal version;
+    private Long version;
 }
